refactor: package lowercase, adviceController 네이밍 등 리뷰 반영
