+++ src/main/kotlin/com/devfloor/untitled/articleoption/application/ArticleOptionService.kt
@@ -1,8 +1,8 @@
-package com.devfloor.untitled.articleOption.application
+package com.devfloor.untitled.articleoption.application
 
 import com.devfloor.untitled.article.domain.Article
-import com.devfloor.untitled.articleOption.domain.ArticleOption
-import com.devfloor.untitled.articleOption.domain.ArticleOptionRepository
+import com.devfloor.untitled.articleoption.domain.ArticleOption
+import com.devfloor.untitled.articleoption.domain.ArticleOptionRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
