+++ src/main/kotlin/com/devfloor/untitled/articleoption/domain/ArticleOption.kt
@@ -1,10 +1,9 @@
-package com.devfloor.untitled.articleHashtag.domain
+package com.devfloor.untitled.articleoption.domain
 
 import com.devfloor.untitled.article.domain.Article
-import com.devfloor.untitled.hashtag.domain.Hashtag
+import com.devfloor.untitled.option.domain.Option
 import javax.persistence.Column
 import javax.persistence.Entity
-import javax.persistence.FetchType
 import javax.persistence.GeneratedValue
 import javax.persistence.GenerationType
 import javax.persistence.Id
@@ -13,14 +12,14 @@ import javax.persistence.ManyToOne
 import javax.persistence.Table
 
 @Entity
-@Table(name = "article_hashtag")
-class ArticleHashtag(
+@Table(name = "article_option")
+class ArticleOption(
     article: Article,
-    hashtag: Hashtag,
+    option: Option
 ) {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "article_hashtag_id")
+    @Column(name = "article_option_id")
     var id: Long? = null
         protected set
 
@@ -29,6 +28,6 @@ class ArticleHashtag(
     val article: Article = article
 
     @ManyToOne
-    @JoinColumn(name = "hashtag_id")
-    val hashtag: Hashtag = hashtag
+    @JoinColumn(name = "option_id")
+    val option: Option = option
 }
