+++ src/main/kotlin/com/devfloor/untitled/common/presentation/CommonAdviceController.kt
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.ControllerAdvice
 import org.springframework.web.bind.annotation.ExceptionHandler
 
 @ControllerAdvice
-class AdviceController {
+class CommonAdviceController {
     @ExceptionHandler(value = [NotFoundException::class])
     fun handleNoSuchElementException(): ResponseEntity<Any> =
         ResponseEntity(HttpStatus.NOT_FOUND)
